+++ rest-server/src/main/java/com/readlearncode/dukesbookshop/restserver/infrastructure/exceptionManagers/AlreadySignedCustomerMapper.java
@@ -1,6 +1,6 @@
 package com.readlearncode.dukesbookshop.restserver.infrastructure.exceptionManagers;
 
-import com.readlearncode.dukesbookshop.restserver.infrastructure.exception.CustomerAlreadySigned;
+import com.readlearncode.dukesbookshop.restserver.infrastructure.exceptions.CustomerAlreadySignedException;
 
 import javax.ws.rs.core.Response;
 import javax.ws.rs.ext.ExceptionMapper;
@@ -10,9 +10,9 @@ import javax.ws.rs.ext.Provider;
  * Created by navid on 11/25/17.
  */
 @Provider
-public class AlreadySignedCustomerMapper implements ExceptionMapper<CustomerAlreadySigned> {
+public class AlreadySignedCustomerMapper implements ExceptionMapper<CustomerAlreadySignedException> {
     @Override
-    public Response toResponse(CustomerAlreadySigned customerAlreadySigned) {
+    public Response toResponse(CustomerAlreadySignedException customerAlreadySignedException) {
         return Response.status(Response.Status.CONFLICT).build();
     }
 }
