+++ rest-server/src/main/java/com/readlearncode/dukesbookshop/restserver/infrastructure/exceptionManagers/ReservationExceptionMapper.java
@@ -1,6 +1,6 @@
 package com.readlearncode.dukesbookshop.restserver.infrastructure.exceptionManagers;
 
-import com.readlearncode.dukesbookshop.restserver.infrastructure.exception.ReservationException;
+import com.readlearncode.dukesbookshop.restserver.infrastructure.exceptions.ReservationException;
 
 import javax.ws.rs.core.Response;
 import javax.ws.rs.ext.ExceptionMapper;
