+++ rest-server/src/main/java/com/readlearncode/dukesbookshop/restserver/infrastructure/exceptions/CustomerAlreadySignedException.java
@@ -1,16 +1,16 @@
-package com.readlearncode.dukesbookshop.restserver.infrastructure.exception;
+package com.readlearncode.dukesbookshop.restserver.infrastructure.exceptions;
 
 /**
  * Created by navid on 11/25/17.
  */
-public class CustomerAlreadySigned extends Exception {
+public class CustomerAlreadySignedException extends Exception {
     private String message;
 
-    public CustomerAlreadySigned() {
+    public CustomerAlreadySignedException() {
         this.message = "Someone has signed with this phone number before!";
     }
 
-    public CustomerAlreadySigned(String message) {
+    public CustomerAlreadySignedException(String message) {
         this.message = message;
     }
