+++ rest-server/src/main/java/com/readlearncode/dukesbookshop/restserver/infrastructure/exceptions/MenuItemNotFoundException.java
@@ -1,4 +1,4 @@
-package com.readlearncode.dukesbookshop.restserver.infrastructure.exception;
+package com.readlearncode.dukesbookshop.restserver.infrastructure.exceptions;
 
 /**
  * Created by navid on 1/10/18.
